[ISSUE-326] 휴리스틱 값 : 거리 - 안전수치로 변경
- 작을 수록 좋음
