+++ src/main/java/com/antybeety/map/way/controller/FindPathController.java
@@ -33,44 +33,44 @@ public class FindPathController {
         return distanceCalc.calcDistance(startLat, startLon, endLat, endLon);
     }
 
-    public double calcSafetyValue(double startLat, double startLon, double endLat, double endLon){
-
-        double ea;  // left
-        double ja;  // right
-        double ka;  // top
-        double la;  // bottom
-
-        double area;    // 영역의 넓이
-
-        Map<String,Object> bounds = new HashMap<>();
-
-        // top, bottom 찾기
-        if(startLat > endLat){
-            la = endLat;
-            ka = startLat;
-        } else {
-            la = startLat;
-            ka = endLat;
-        }
-
-        // left, right 찾기
-        if(startLon > endLon){
-            ea = endLon;
-            ja = startLon;
-        }else {
-            ea = startLon;
-            ja = endLon;
-        }
-
-        // 0.00005를 더해주는 이유 : 정확히 그 내부가 아닌 도로의 폭 정도를 여유를 주기 위함
-        bounds.put("la",la-0.00005);
-        bounds.put("ka",ka+0.00005);
-        bounds.put("ja",ja+0.00005);
-        bounds.put("ea",ea-0.00005);
-
-
-        return safetyValue.calcSafetyValue(bounds);
-    }
+//    public double calcSafetyValue(double startLat, double startLon, double endLat, double endLon){
+//
+//        double ea;  // left
+//        double ja;  // right
+//        double ka;  // top
+//        double la;  // bottom
+//
+//        double area;    // 영역의 넓이
+//
+//        Map<String,Object> bounds = new HashMap<>();
+//
+//        // top, bottom 찾기
+//        if(startLat > endLat){
+//            la = endLat;
+//            ka = startLat;
+//        } else {
+//            la = startLat;
+//            ka = endLat;
+//        }
+//
+//        // left, right 찾기
+//        if(startLon > endLon){
+//            ea = endLon;
+//            ja = startLon;
+//        }else {
+//            ea = startLon;
+//            ja = endLon;
+//        }
+//
+//        // 0.00005를 더해주는 이유 : 정확히 그 내부가 아닌 도로의 폭 정도를 여유를 주기 위함
+//        bounds.put("la",la-0.00005);
+//        bounds.put("ka",ka+0.00005);
+//        bounds.put("ja",ja+0.00005);
+//        bounds.put("ea",ea-0.00005);
+//
+//
+//        return safetyValue.calcSafetyValue(bounds);
+//    }
 
     public List<NodeVO> searchSafePath(double startLat, double startLon, double endLat, double endLon) {
